added updates after help call, done_items not showing
