+++ todo_app/templates/index.html
@@ -23,7 +23,7 @@
 		<table>
 		<h3><b>To do</h3></b>    
 		                                                                                                   
-		{%for item in view_model.items : %}
+		{%for item in view_model.todolist : %}
 		<div>
 				<br>
 				
@@ -45,7 +45,7 @@
 			</table>
 			<table>
 		<br><h3><b> Doing </h3></b>  
-{%for doing in doing_objects: %}
+{%for doing in view_model.doinglist: %}
 		<div>
 				<br>
 				
@@ -55,18 +55,28 @@
 			<table id ="done-table"> 
 
 			<br><h3><b>Done<br></h3>
-<tr>{%for done in done_objects: %}<tr>   <br>
-				<tr>{{done.name}}<div></tr> 
+
+		   {% if view_model.show_all_done_items():   %}
+		  		 {
+					{% for done in view_model.donelist: %}   
+					<tr>{{done.name}}<div></tr> 
+
 				 <form action="/update_back" method="post">
-           <input type="hidden" name="n" value="{{done["id"]}}">
-           <input type="submit" value="Return to To do">
-           </form>	
-		   <!--<form action="/toggle_view" method="post">
-			<input type="hidden" name="i" value="{{done["id"]}}">
-			<input type="submit" value="Toggle">-->
-			</form>	
+          		 <input type="hidden" name="n" value="{{done["id"]}}">
+           		 <input type="submit" value="Return to To do">
+           			</form>
+					   {% endfor %}	
+		   }
+		    {% else  %}
+		   <details>
+			   {{view_model.older_done_items()}}
+			<summary>
+				{{view_model.recent_done_items()}}
+			</summary>
+		   </details>
+		   {% endif %}
 		   </div>
-{% endfor %}<br>			
+			
 		
    </table>
